# openapi/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ApiResponseConsultingLeadResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/ConsultingLeadResponse"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorResponse"
          },
          "success" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseIssued" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/Issued"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorResponse"
          },
          "success" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ApiResponseIssuedDetail" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/IssuedDetail"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorResponse"
          },
          "success" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseListIssuedSummary" : {
        "properties" : {
          "data" : {
            "items" : {
              "$ref" : "#/components/schemas/IssuedSummary"
            },
            "type" : "array"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorResponse"
          },
          "success" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseUploaded" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/Uploaded"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorResponse"
          },
          "success" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseVoid" : {
        "properties" : {
          "data" : {
            "type" : "object"
          },
          "error" : {
            "$ref" : "#/components/schemas/ErrorResponse"
          },
          "success" : {
            "type" : "boolean"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ConsultingLeadResponse" : {
        "properties" : {
          "diagnosisId" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "maskedEmail" : {
            "type" : "string"
          },
          "maskedPhone" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "DataBuffer" : {
        "type" : "object"
      },
...
      "ErrorResponse" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "details" : {
            "additionalProperties" : {
              "type" : "object"
            },
            "type" : "object"
          },
          "fieldErrors" : {
            "items" : {
              "$ref" : "#/components/schemas/FieldError"
            },
            "type" : "array"
          },
          "message" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "FieldError" : {
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "rejectedValue" : {
            "type" : "object"
          }
        },
        "type" : "object"
      },
...
      "Issue" : {
        "properties" : {
          "templateCode" : {
            "minLength" : 1,
            "type" : "string"
          },
          "values" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "type" : "object"
          }
        },
        "type" : "object"
      },
      "Issued" : {
        "properties" : {
          "documentId" : {
            "type" : "string"
          },
          "downloadUrl" : {
            "type" : "string"
          },
          "issuedAt" : {
            "type" : "string"
          },
          "templateCode" : {
            "type" : "string"
          },
          "templateName" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "IssuedDetail" : {
        "properties" : {
          "documentId" : {
            "type" : "string"
          },
          "downloadUrl" : {
            "type" : "string"
          },
          "issuedAt" : {
            "type" : "string"
          },
          "templateCode" : {
            "type" : "string"
          },
          "templateName" : {
            "type" : "string"
          },
          "values" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "type" : "object"
          }
        },
        "type" : "object"
      },
      "IssuedSummary" : {
        "properties" : {
          "documentId" : {
            "type" : "string"
          },
          "downloadUrl" : {
            "type" : "string"
          },
          "issuedAt" : {
            "type" : "string"
          },
          "templateCode" : {
            "type" : "string"
          },
          "templateName" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "RequestConsultingRequest" : {
        "properties" : {
          "consentVersion" : {
            "minLength" : 1,
            "type" : "string"
          },
          "contactEmail" : {
            "type" : "string"
          },
          "contactName" : {
            "minLength" : 1,
            "type" : "string"
          },
          "contactPhone" : {
            "minLength" : 1,
            "type" : "string"
          },
          "diagnosisId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "message" : {
            "type" : "string"
          },
          "privacyConsent" : {
            "type" : "boolean"
          },
          "sensitiveInfoConsent" : {
            "type" : "boolean"
          },
          "serviceLimitAcknowledged" : {
            "type" : "boolean"
          }
        },
        "required" : [ "diagnosisId" ],
        "type" : "object"
      },
...
      "Uploaded" : {
        "properties" : {
          "contentType" : {
            "type" : "string"
          },
          "downloadUrl" : {
            "type" : "string"
          },
          "fileId" : {
            "type" : "string"
          },
          "originalName" : {
            "type" : "string"
          },
          "sizeInBytes" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "로그인(POST /api/v1/auth/login) 응답의 accessToken",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "1인 제조기업·소공인 제품 인증 준비도 자가진단 및 컨설팅 연계 서비스.\n\n• 모든 응답은 ApiResponse 봉투(success/data/error/traceId/timestamp)로 감쌉니다.\n• 준비도 점수는 인증 합격·불합격 판정이 아니라 준비 상태를 확인하는 사전 점검 지표입니다.\n• 인증이 필요한 엔드포인트는 우상단 Authorize에 로그인 accessToken을 넣고 호출하세요.\n",
    "license" : {
      "name" : "팀 육성사이다"
    },
    "title" : "OpenCertFlow API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/consulting-leads" : {
      "post" : {
        "operationId" : "request",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RequestConsultingRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseConsultingLeadResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "상담 신청(공개)" ]
      }
    },
...
    "/api/v1/documents/issues" : {
      "get" : {
        "operationId" : "myDocuments",
        "parameters" : [ {
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "format" : "int32",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListIssuedSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "document-controller" ]
      },
      "post" : {
        "operationId" : "issue",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Issue"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseIssued"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "document-controller" ]
      }
    },
    "/api/v1/documents/issues/{documentId}" : {
      "get" : {
        "operationId" : "get_5",
        "parameters" : [ {
          "in" : "path",
          "name" : "documentId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseIssuedDetail"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "document-controller" ]
      }
    },
...
    "/api/v1/files" : {
      "post" : {
        "operationId" : "upload",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "properties" : {
                  "file" : {
                    "format" : "binary",
                    "type" : "string"
                  }
                },
                "required" : [ "file" ],
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseUploaded"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "file-controller" ]
      }
    },
    "/api/v1/files/{id}" : {
      "delete" : {
        "operationId" : "delete_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "file-controller" ]
      },
      "get" : {
        "operationId" : "download",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/DataBuffer"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "file-controller" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "description" : "Generated server url",
    "url" : ""
  } ],
...
}